100 % test case coverage
  py38: OK (15.37=setup[7.19]+cmd[6.77,0.32,1.09] seconds)
  py39: OK (11.49=setup[3.81]+cmd[6.39,0.39,0.90] seconds)
  py310: OK (11.70=setup[3.71]+cmd[6.75,0.32,0.92] seconds)
  py311: OK (10.22=setup[2.32]+cmd[6.00,0.51,1.39] seconds)
  congratulations :) (48.90 seconds)

Notes.

This commit has one outstanding issue:

Python 3.8 and 3.9 do not execute line 103 of production.py.
Python 3.10 and 3.11 do execute this line.  This line was added *because* it seemed like Ncc conditions where being silently dropped.

The isinstance() matching seems to have changed with python 3.10 and the introduction of the pipe operator for Union types.